[ci skip] Update TaskAgent Tests - Fix type for subscripting array

+++ Tests/UnitTests/Tasks/TaskAgentTests.cpp
@@ -179,7 +179,7 @@ TEST(TaskAgent, RunMultiTaskWithBrief)
     };
 
     traits.reserve(5);
-    for (int i = 0; i < 5; ++i)
+    for (size_t i = 0; i < 5; ++i)
     {
         traits.emplace_back(TestUtils::GenerateRandomTrait());
         if (traits[i].id == +TaskID::BRIEF)
